MAJOR UPDATE (read description)
The virtual keyboard is now a FLOATING WIDGET ; not a panel anymore. This allows moving the keyboard where you want on screen and allows using it on SmartPanels such as the App Launcher.
A few issues remain :
I'd like the keyboard to be scaleable in real time via drag n' drop.
The keyboard can't be dragged from a screen to another.